Update the notification microservice to consume the payload from the appropriate queue
When the notification microservice is up and running, it listens for any messages in the queue specified.

Once at least one message is in the queue, the microservice consumes the message payload and saves a new notification object constructed from the payload into the database.

Also, set up and configure the exchange and queue for RabbitMQ broker when the customer microservice publishes the payload to appropriate exchange.