Add the initial files for the clients child module
This module initially contains the feign client for the Album service.

The clients child module is responsible for containing the HTTP clients using the openfeign dependency to other internal microservices or external clients to 3rd party services.

As a result, other microservice modules can utilize these clients to communicate with other microservices or 3rd party services.