Update the User Service to include a new endpoint for retrieving the user's details
The new GET endpoint for retrieving the user's details utilizes the albumClient.getUserAlbums(userId) to get a list of albums for a user by making a synchronous HTTP request to the Album Service.

In addition, enable HTTP request logging for the Album Client and handle response errors made from the Album Client with FeignErrorDecoder.